Add user signup error handling and implement password reset code functionality
