+++ backend/app/routers/authentication/authenticate.py
@@ -1,11 +1,8 @@
 # pylint: disable=no-value-for-parameter
 
 import random
-import re
-from datetime import datetime, timedelta, timezone
+from datetime import datetime
 
-import bcrypt
-import jwt
 from fastapi import Depends, HTTPException, status
 from fastapi.security import OAuth2PasswordBearer
 from snowflake import SnowflakeGenerator
@@ -14,22 +11,28 @@ from app.data_layer.database.crud.user_crud import (
     create_user,
     get_user,
     get_user_by_attr,
-    is_attr_data_in_db,
     update_user,
 )
 from app.data_layer.database.models.user_model import Gender, User
 from app.schemas.user_model import UserSignup
+from app.utils.common.exceptions.authentication import UserSignupError
 from app.utils.constants import (
     ACCESS_TOKEN_EXPIRE_MINUTES,
-    JWT_HASHING_ALGO,
+    EMAIL,
     JWT_REFRESH_SECRET,
     JWT_SECRET,
     MACHINE_ID,
     REFRESH_TOKEN_EXPIRE_DAYS,
+    USER_ID,
 )
 
-USER_ID = "user_id"
-EMAIL = "email"
+from .jwt_tokens import create_token, decode_token
+from .user_validation import (
+    get_hash_password,
+    validate_user_data,
+    validate_user_exists,
+    verify_password,
+)
 
 snowflake_generator = SnowflakeGenerator(MACHINE_ID)
 
@@ -46,261 +49,6 @@ def get_snowflake_id() -> int:
         The unique snowflake ID
     """
     return next(snowflake_generator)
-
-
-class UserSignupError(HTTPException):
-    """
-    Exception class for user signup errors. This class is used to raise exceptions
-    when there is an error during user signup. This exception raises when there is
-    invalid user data.
-
-    Attributes:
-    -----------
-    message: ``str``
-        The error message to be displayed to the user
-    """
-
-    def __init__(self, message: str):
-        super().__init__(status.HTTP_400_BAD_REQUEST, message)
-
-
-def validate_email(email: str):
-    """
-    This function validates the email format. It allows only email addresses with
-    the domain 'gmail.com'
-
-    Parameters:
-    -----------
-    email: ``str``
-        The email address to be validated
-    """
-    email_regex = r"^[\w\.-]+@gmail\.com$"
-
-    if re.match(email_regex, email) is None:
-        raise UserSignupError("Invalid email format")
-
-
-def validate_phone_number(phone_number: str) -> None:
-    """
-    This function validates the phone number format. It allows only phone numbers
-    with 10 digits.
-
-    Parameters:
-    -----------
-    phone_number: ``str``
-        The phone number to be validated
-    """
-    phone_number_regex = r"\d{10}$"
-
-    if re.fullmatch(phone_number_regex, phone_number) is None:
-        raise UserSignupError("Invalid phone number format")
-
-
-def get_hash_password(password: str) -> str:
-    """
-    It hashes the password using bcrypt and returns the hashed password.
-
-    Parameters:
-    -----------
-    password: ``str``
-        The password to be hashed
-
-    Returns:
-    --------
-    ``str``
-        The hashed password
-    """
-    return bcrypt.hashpw(password.encode("utf-8"), bcrypt.gensalt()).decode("utf-8")
-
-
-def validate_password(password: str):
-    """
-    Validates the password format. The password must be at least 8 characters long
-    and include an uppercase letter, lowercase letter, digit, and special character.
-
-    Parameters:
-    -----------
-    password: ``str``
-        The password to be validated
-    """
-
-    if not (
-        len(password) >= 8
-        and re.search(r"[A-Z]", password) is not None
-        and re.search(r"[a-z]", password) is not None
-        and re.search(r"\d+", password) is not None
-        and re.search(r"[!@#$%^&*(),.?\":{}|<>]", password) is not None
-    ):
-
-        raise UserSignupError(
-            "Password must be at least 8 characters long and include an uppercase letter, "
-            "lowercase letter, digit, and special character",
-        )
-
-
-def verify_password(password: str, hash_password: str) -> None:
-    """
-    Verifies the password by comparing the hashed password with the password.
-
-    Parameters:
-    -----------
-    password: ``str``
-        The password to be verified
-    hash_password: ``str``
-        The hashed password to be compared with the password
-    """
-    if not bcrypt.checkpw(password.encode("utf-8"), hash_password.encode("utf-8")):
-        raise HTTPException(status.HTTP_401_UNAUTHORIZED, "Passwords do not match")
-
-
-def validate_date_of_birth(date_of_birth: str) -> None:
-    """
-    Validates the date of birth format. The date of birth must be in the format
-    'dd/mm/yyyy' and cannot be in the future.
-
-    Parameters:
-    -----------
-    date_of_birth: ``str``
-        The date of birth to be validated
-    """
-    try:
-        dob = datetime.strptime(date_of_birth, "%d/%m/%Y")
-
-        if dob >= datetime.now():
-            raise UserSignupError("Date of birth cannot be in the future")
-
-    except ValueError as exc:
-        raise UserSignupError(
-            "Invalid date format for date of birth. Expected format: dd/mm/yyyy",
-        ) from exc
-
-
-def validate_user_exists(user: UserSignup) -> str | None:
-    """
-    Checks if the user already exists in the database. It checks the email
-    and phone number fields to see if the user already exists.
-
-    Parameters:
-    -----------
-    user: ``UserSignup``
-        The user details to be checked
-
-    Returns:
-    --------
-    ``str | None``
-        An error message if the user already exists, otherwise None
-    """
-    fields_to_check = {
-        "email": user.email,
-        "phone_number": user.phone_number,
-    }
-    response = is_attr_data_in_db(User, fields_to_check)
-
-    return response
-
-
-def validate_user_data(user: UserSignup) -> None:
-    """
-    Validates the user data. If any of the user data is invalid, it raises an exception.
-
-    Parameters:
-    -----------
-    user: ``UserSignup``
-        The user details to be validated
-    """
-    if user.password != user.confirm_password:
-        raise UserSignupError("Passwords do not match")
-
-    validate_email(user.email)
-    validate_phone_number(user.phone_number)
-    validate_password(user.password)
-    validate_date_of_birth(user.date_of_birth)
-
-
-def create_token(data: dict, secret: str, expire_time: float) -> str:
-    """
-    Creates a JWT token with the given data and expiration time.
-
-    Parameters:
-    -----------
-    data: ``dict``
-        The data to be encoded in the token
-    secret: ``str``
-        The secret key used to encode the token
-    expire_time: ``int``
-        The expiration time of the token in minutes
-
-    Returns:
-    --------
-    ``str``
-        The JWT token
-    """
-    to_encode = data.copy()
-
-    # Using `exp` as token expiry time, if `exp` is present in the payload data then
-    # pyJwt will verify the token expiry automatically
-    to_encode.update(
-        {"exp": datetime.now(timezone.utc) + timedelta(minutes=expire_time)}
-    )
-
-    return jwt.encode(to_encode, secret, algorithm=JWT_HASHING_ALGO)
-
-
-def decode_token(token: str, secret: str) -> dict[str, str]:
-    """
-    Decodes the given token using the secret key and returns the decoded data.
-    The decoded data contain the payload used for creating the token.
-
-    Parameters:
-    -----------
-    token: ``str``
-        The token to be decoded
-    secret: ``str``
-        The secret key used to decode the token
-
-    Returns:
-    --------
-    ``dict[str, str]``
-        The decoded data from the token
-    """
-    try:
-        return jwt.decode(token, secret, algorithms=[JWT_HASHING_ALGO])
-    except jwt.ExpiredSignatureError as exc:
-        raise HTTPException(status.HTTP_401_UNAUTHORIZED, "Token has expired") from exc
-    except jwt.InvalidTokenError as exc:
-        raise HTTPException(status.HTTP_401_UNAUTHORIZED, "Invalid token") from exc
-
-
-def access_token_from_refresh_token(refresh_token: str) -> dict[str, str]:
-    """
-    Create the access token using the refresh token. If the refresh token is invalid
-    or expired, it returns an error message. Otherwise, it generates a new access token
-    and returns it.
-
-    Parameters:
-    -----------
-    refresh_token: ``str``
-        The refresh token to generate a new access token
-
-    Returns:
-    --------
-    ``dict[str, str]``
-        A dictionary containing the new access token and the refresh token
-    """
-    decoded_data = decode_token(refresh_token, JWT_REFRESH_SECRET)
-
-    get_user(decoded_data[USER_ID])
-
-    access_token = create_token(
-        {USER_ID: decoded_data[USER_ID], EMAIL: decoded_data[EMAIL]},
-        JWT_SECRET,
-        ACCESS_TOKEN_EXPIRE_MINUTES,
-    )
-
-    return {
-        "access_token": access_token,
-        "refresh_token": refresh_token,
-    }
 
 
 def signup_user(user: UserSignup):
