+++ backend/app/utils/common/exceptions/historical_data.py
