+++ backend/tests/utils/test_date_utils.py
@@ -2,7 +2,7 @@ from datetime import datetime
 
 import pytest
 
-from app.utils.common.exceptions import InvalidDateTimeFormatException
+from app.utils.common.exceptions.historical_data import InvalidDateTimeFormatException
 from app.utils.date_utils import validate_datetime_format
